content: rename router to hub
